MessageBalanceImplementation

+++ SMSProposal/DataServiceLibrary/ITemplateService.cs
@@ -12,7 +12,7 @@ namespace DataServiceLibrary
         Task<ICollection<TemplateViewModel>> GetPagedTemplates(int subcriberId, int skip, int pagesize, string ordercolumn, bool desc);
         Task<ICollection<TemplateViewModel>> GetTemplates(int subcriberId);
         Task<int> TotalTemplates(int subcriberId);
-        Task<SubscriberTemplate> FindTemplate(int subcriberId,int templateId=0, string templatename=null);
+        Task<SubscriberTemplate> FindTemplate(int subcriberId,long templateId=0, string templatename=null);
         Task<SubscriberTemplate> AddTemplate(TemplateViewModel templateViewModel,int subscirberId);
         Task<int> SaveAsync();
     }
